+++ packages/shared/src/parsers/rss.parser.ts
@@ -3,8 +3,9 @@ import {parseZodResult} from '@shared/lib/parser.shared';
 import {makeSuccessResult} from '@shared/lib/results.shared';
 
 import type {Result} from '@shared/types/results.types';
-import {RssFeedProviderTypeSchema} from '@shared/types/rss.types';
 import type {RssFeedProviderType} from '@shared/types/rss.types';
+
+import {RssFeedProviderTypeSchema} from '@shared/schemas/rss.schema';
 
 /**
  * Parses a {@link RssFeedProviderType} from a string. Returns an `ErrorResult` if the string is not
