+++ packages/shared/src/schemas/eventLog.schema.ts
@@ -2,11 +2,11 @@ import {z} from 'zod';
 
 import {Environment, EventType} from '@shared/types/eventLog.types';
 import {FeedItemActionType} from '@shared/types/feedItems.types';
-import {FirestoreTimestampSchema} from '@shared/types/firebase.types';
 
 import {AccountIdSchema} from '@shared/schemas/accounts.schema';
 import {ActorSchema} from '@shared/schemas/actors.schema';
 import {FeedItemIdSchema} from '@shared/schemas/feedItems.schema';
+import {FirestoreTimestampSchema} from '@shared/schemas/firebase.schema';
 import {UserFeedSubscriptionIdSchema} from '@shared/schemas/userFeedSubscriptions.schema';
 
 /** Zod schema for an {@link EventId}. */
