+++ packages/shared/src/schemas/rss.schema.ts
@@ -1,0 +1,3 @@
+import {z} from 'zod';
+
+export const RssFeedProviderTypeSchema = z.union([z.literal('local'), z.literal('superfeedr')]);
