+++ packages/shared/src/types/rss.types.ts
@@ -1,9 +1,6 @@
 import type {AsyncResult, Result} from '@conductor/shared/src/types/results.types';
-import {z} from 'zod';
 
 export type RssFeedProviderType = 'local' | 'superfeedr';
-
-export const RssFeedProviderTypeSchema = z.union([z.literal('local'), z.literal('superfeedr')]);
 
 /** Provides a way to subscribe to RSS feeds and be notified of new items. */
 export interface RssFeedProvider {
