+++ packages/shared/src/types/superfeedr.types.ts
@@ -1,81 +1,7 @@
-import {z} from 'zod';
-
-import {prefixResultIfError} from '@shared/lib/errorUtils.shared';
-import {parseZodResult} from '@shared/lib/parser.shared';
-
-import type {Result} from '@shared/types/results.types';
-
 /**
  * Internal type for the credentials needed to use the Superfeedr API.
  */
 export interface SuperfeedrCredentials {
   readonly user: string;
   readonly apiKey: string;
-}
-
-const SuperfeedrActorSchema = z.object({
-  displayName: z.string(),
-  id: z.string(),
-});
-
-const SuperfeedrStandardLinkSchema = z.object({
-  title: z.string(),
-  href: z.string(),
-  rel: z.string(),
-  type: z.string().optional(),
-});
-
-const SuperfeedrFeedSourceSchema = z.object({
-  id: z.string(),
-  title: z.string(),
-  updated: z.number(),
-  published: z.number(),
-  permalinkUrl: z.string().url(),
-  standardLinks: z.object({
-    alternate: z.array(SuperfeedrStandardLinkSchema),
-    superfeedr: z.array(SuperfeedrStandardLinkSchema),
-  }),
-});
-
-const SuperfeedrFeedItemSchema = z.object({
-  id: z.string(),
-  published: z.number(),
-  updated: z.number(),
-  title: z.string(),
-  summary: z.string(),
-  permalinkUrl: z.string().url(),
-  standardLinks: z.object({
-    alternate: z.array(SuperfeedrStandardLinkSchema),
-  }),
-  actor: SuperfeedrActorSchema,
-  source: SuperfeedrFeedSourceSchema,
-});
-
-const SuperfeedWebhookRequestBodySchema = z.object({
-  status: z.object({
-    code: z.number(),
-    http: z.string(),
-    nextFetch: z.number(),
-    velocity: z.number(),
-    bozoRank: z.number(),
-    title: z.string(),
-    period: z.number(),
-    lastFetch: z.number(),
-    lastParse: z.number(),
-    lastMaintenanceAt: z.number(),
-    feed: z.string(),
-  }),
-  title: z.string(),
-  updated: z.number().nullable(),
-  id: z.string(),
-  items: z.array(SuperfeedrFeedItemSchema),
-});
-
-type SuperfeedWebhookRequestBody = z.infer<typeof SuperfeedWebhookRequestBodySchema>;
-
-export function parseSuperfeedrWebhookRequestBody(
-  maybeBody: unknown
-): Result<SuperfeedWebhookRequestBody> {
-  const parsedResult = parseZodResult(SuperfeedWebhookRequestBodySchema, maybeBody);
-  return prefixResultIfError(parsedResult, 'Unexpected Superfeedr webhook request');
 }
